+++ src/mesido/workflows/goals/minimize_tco_goal.py
@@ -108,6 +108,9 @@ class MinimizeTCO(Goal):
         If `discounted_annualized_cost` is not defined or False, the operational and
         fixed operational costs are multiplied by the number_of_years.
 
+        'technicalLifetime' is function of installation and investment cost calculations
+
+
         Args:
             optimization_problem (TechnoEconomicMixin): The optimization problem instance.
             cost_type (str): The type of cost to calculate ("operational",
@@ -122,6 +125,12 @@ class MinimizeTCO(Goal):
         obj = 0.0
         for asset_type in asset_types:
             for asset in optimization_problem.energy_system_components.get(asset_type, []):
+                if asset != optimization_problem.energy_system_components.get("heat_demand", []):
+                    technical_lifetime = self.number_of_years
+                else:
+                    technical_lifetime = optimization_problem.parameters(0)[
+                        f"{asset}.technical_life"
+                    ]
                 extra_var = optimization_problem.extra_variable(cost_type_map[asset])
                 if options["discounted_annualized_cost"]:
                     # We only want the operational cost for a single year when we use
@@ -131,7 +140,7 @@ class MinimizeTCO(Goal):
                     obj += extra_var * self.number_of_years
                 else:
                     # These are the CAPEX cost under non-annualized condition
-                    obj += extra_var
+                    obj += extra_var * (self.number_of_years / technical_lifetime)
         return obj
 
     def function(self, optimization_problem: TechnoEconomicMixin, ensemble_member) -> MX: