+++ tests/test_end_scenario_sizing.py
@@ -100,12 +100,20 @@ class TestEndScenarioSizing(TestCase):
             *self.solution.energy_system_components.get("heat_pump", []),
             *self.solution.energy_system_components.get("heat_pipe", []),
         ]:
+            if asset != self.solution.energy_system_components.get("heat_demand", []):
+                technical_lifetime = years
+            else:
+                technical_lifetime = self.solution.parameters(0)[f"{asset}.technical_life"]
             obj += self.results[f"{self.solution._asset_fixed_operational_cost_map[asset]}"] * years
             obj += (
                 self.results[f"{self.solution._asset_variable_operational_cost_map[asset]}"] * years
             )
-            obj += self.results[f"{self.solution._asset_investment_cost_map[asset]}"]
-            obj += self.results[f"{self.solution._asset_installation_cost_map[asset]}"]
+            obj += self.results[f"{self.solution._asset_investment_cost_map[asset]}"] * (
+                years / technical_lifetime
+            )
+            obj += self.results[f"{self.solution._asset_installation_cost_map[asset]}"] * (
+                years / technical_lifetime
+            )
 
         np.testing.assert_allclose(obj / 1.0e6, self.solution.objective_value)
 
@@ -128,6 +136,8 @@ class TestEndScenarioSizing(TestCase):
         - Unstaged approaches, using the general function run_optimization_problem and the
         function run_end_scenario_sizing with staged_pipe_optimization to False should have
         comparable computation times.
+
+        - Check if TCO goal included the desired cost components
 
         Missing:
         - Link ATES t0 utilization to state of charge at end of year for optimizations over one
@@ -190,10 +200,18 @@ class TestEndScenarioSizing(TestCase):
             *solution_staged.energy_system_components.get("heat_pump", []),
             *solution_staged.energy_system_components.get("heat_pipe", []),
         ]:
+            if asset != solution_staged.energy_system_components.get("heat_demand", []):
+                technical_lifetime = years
+            else:
+                technical_lifetime = solution_staged.parameters(0)[f"{asset}.technical_life"]
             obj += results[f"{solution_staged._asset_fixed_operational_cost_map[asset]}"] * years
             obj += results[f"{solution_staged._asset_variable_operational_cost_map[asset]}"] * years
-            obj += results[f"{solution_staged._asset_investment_cost_map[asset]}"]
-            obj += results[f"{solution_staged._asset_installation_cost_map[asset]}"]
+            obj += results[f"{solution_staged._asset_investment_cost_map[asset]}"] * (
+                years / technical_lifetime
+            )
+            obj += results[f"{solution_staged._asset_installation_cost_map[asset]}"] * (
+                years / technical_lifetime
+            )
 
         np.testing.assert_allclose(obj / 1.0e6, solution_staged.objective_value)